Add area-type selection radio buttons
Fetches filter job information and determines if the current dimension
is an area type. If so, we fetch a list of area types from the
dimensions API and display them on the page.

Currently the dimension name (which the importer maps to a human
readable name) is used as the value for consistency. We'll probably want
to revisit this soon, though, as matching up an encoded name in the URL
is less than ideal.

Resolves: 5569 (https://trello.com/c/yOjZDYAj)