Hook up Change button on area dimension screen
Hooks up the button underneath the area-type radio selection screen so
that it updates the dimension in Mongo (via the filter-flex API).

We still have the ID/Label mismatch with the imported Cantabular
data/dataset dimensions returned from the dataset API
(https://github.com/ONSdigital/dp-frontend-filter-flex-dataset/pull/12#discussion_r835367062).
Until this is fixed, I've changed the radio button value to be the
pretty-name rather than the ID. Without this, the dimension is never
found and the journey doens't work. Hopefully once this is fixed we can
just swap back to using the ID.

Resolves: 5643 (https://trello.com/c/bIwXXHiR)