Fix radio type page post filter model update
A recent change to the filter model has broken the area-type selection
screen (as well as the screen for other dimensions).

This re-instates the previous behaviour using a different method, namely
fetching the dimension via the API client, rather than using the
dimensions on the filter job (which are no longer present).

The name/ID/label shennangians will be revistied shortly and then we can
get rid of it.

Resolves: 5659 (https://trello.com/c/5jX64Vhm)