Add client-facing validation to area selection
A user shouldn't really ever end up in a position where they don't have
a preselected area-type already, but on the off chance they do this
commit ensures that when we validate incoming requests, if the
`dimension` field is missing we show an error on the UI (rather than
just 400'ing).

Hidden fields still 400, as there's nothing the user can do to rectify
it - if there's an issue they've either tampered with it or we've broken
something.

Resolves: 5643