Update to the latest dp-api-cients-go
The dp-api-clients-go has one change where the AreaType JSON
representation has been changed to a correct `snake_case` instead of
`kebab-case`, this was changed on the PopulationAPI as well.

At the moment, when we click on the area-types link we do not see all
the area-types beign displayed on the template.

This change should fix it.